moved serializer into same package as serialized class
